requirements.txt
pip
setuptools
chainer
pillow
tensorflow-gpu
